+++ doc/whats-new.rst
@@ -3,8 +3,8 @@
 What's New
 ----------
 
-v0.5.3 (unreleased)
-===================
+v0.6.0 (June 29, 2021)
+======================
 - Support indexing by ``cf_role`` attribute. By `Deepak Cherian`_.
 - Implemented :py:meth:`Dataset.cf.add_canonical_attributes` and :py:meth:`DataArray.cf.add_canonical_attributes`
   to add CF canonical attributes. By `Mattia Almansi`_.